refactor(api): move temporary file utils in test-utils/file.js
